[keyserver] Add thumbhash to upload endpoints
Summary:
This adds possibility to store thumbhash on the keyserver. The thumbhash is saved to the `extra` column of the `uploads` table. Despite a deditated column being discussed, I decided to use the extras one for the sake of simplicity (avoid migration), and also because there are few other fields to consider when redesigning the uploads table schema (e.g. blob holder or encryption key which are even more important than the thumbhash).

Depends on D7802

Test Plan: Added the thumbhash property to the upload multimedia endpoint call in input-state-container. Checked with TablePlus that it's actually saved to the database.

Reviewers: atul, ashoat

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7803