[keyserver] Handle web version in hasMinCodeVersion
Summary: Now that web will have a version we need to handle it in `hasMinCodeVersion`. It now takes two version, one for native and one for web. If either of them is absent we treat it like all versions of the respective platform are allowed.

Test Plan: `yarn flow-all` and check if web app works fine.

Reviewers: kamil, tomek

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7854