[native] Introduce designSystemColors to colors.js
Summary:
Following discussion offline [and in D7740](https://phab.comm.dev/D7740#inline-51019), we want to make `native/themes/colors.js` look more like `web/theme.css`, where the design system colors are defined at the top of the page in their own struct, and the domain-specific usages all references the design system colors.

There are many cases of colors that are using in the mobile app that are NOT in the design system. We'll sync with @ted on this and come back to it later.

Test Plan: Flow, careful reading

Reviewers: atul, tomek, inka, ginsu

Subscribers: