[landing] introduce new header elements to header component
Summary:
This diff is the second part to updating the `Header` component to match the redesigns. The foucs of this diff is to introduce and layout the new elements needed for the header redesign. The new elements include:

  # Beta Badge
  # Team and Investors page nav replacing the App page nav

In this diff I also updated any typography/colors to use our design system created by Ted. I did not update the Github/Twitter colors, because those are one off instances that probably won't ever need to change, but if we want to create variables for them, I can do that.

Depends on D7835

Test Plan:
Please watch the demo video to see how the new `Header` component looks and behaves

iPhone:
{F544670}

iPad:
{F544671}

Desktop:
{F544672}

Responsive:
{F544673}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ted, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7839