[landing] implement on click functions in header
Summary:
This diff is the third part to updating the `Header` component to match the redesigns. The focus of this part is to implement the onClick/tap functions to open and close the mobile nav header. The next diff will address the last part of updating the `Header` which will be to either show an "X" icon or hamburger menu icon in the header depending on if the mobile nav header is active or not

Depends on D7863

Test Plan:
Please watch the demo videos to see how the `Header` looks and behaves:

{F545237}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7868