[web] Added FocusTrap for modals
Summary: We want to prevent users from focusing outside the modals. Added `focus-trap-react` library, to do that, and covered the modals that we use. Added `fallbackFocus` option, which sets focus to div when there are no focusable elements inside the FocusTrap.

Test Plan:
Checked if in edit mode and in other modals user can't focus outside the modal using the tab.
Checked if the modals without focusable elements (i.g. in investors tab
on landing) also work.

Reviewers: michal, inka, ashoat

Subscribers: atul, tomek, ashoat

Differential Revision: https://phab.comm.dev/D7960