[native] Extract function for modyfing items to be displayed in MessageResult
Summary:
issue: https://linear.app/comm/issue/ENG-3864/update-messageresult-style#comment-fe30cc85
For pinned messages and search results to be displayed correctly in the MessageResult, their data needs to be changed. I'm extracting this logic and using it ine MessageResult iteself to avoid duplicating code

Test Plan: Pinned a message, navigated to the pinned messages screen. Checked that the messages display correctly in  the pinned messages screen and in the pin messages modal.

Reviewers: rohan, kamil, bartek, kuba, ashoat

Reviewed By: ashoat

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7770