[native] Add a header componnent with a search bar
Summary:
issue: https://linear.app/comm/issue/ENG-3405/create-a-header-with-the-search-input-field
This will be used in the message search screen. Some design changes were discussed by me and Ted, and noted on Linear.

iOS:
{F532920}
Android:
{F532913}

The search button:
Android:
{F533071}

iOS (my simulator for some reason refuses to show me the keyboard in english, but "szukaj" is "search" in polish):
{F533073}

Test Plan: Tested with next diffs - tested that the query from the search context is set when the "enter" is pressed, and is not being set on every character change. Tested that the cancel button appears only when the input is not empty. Tested that pressing the cancel button clears the query and the input.

Reviewers: kuba, kamil, bartek, ashoat

Reviewed By: ashoat

Subscribers: ted, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7740