[natvie] Add SearchFooter component
Summary:
issue: https://linear.app/comm/issue/ENG-3842/add-the-end-of-results-and-no-results-messages
On the search screen there are different messages that appear at the end of the results (like "End of results") or instead of them (like "No results, ..."). They are all styled the same, so I created a component for them

Test Plan: Displayed the coponent with some message. Checked that it looks as it should.

Reviewers: kuba, kamil, bartek

Reviewed By: kamil

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7772