[native] Add MessageSearchContent
Summary:
Issue: https://linear.app/comm/issue/ENG-3163/create-the-search-screen-for-message-searching
Add a component that will be displayed in the search screen when there is a query typed into the search bar. The logic is pretty similar to pinned messages : D7673
Some navigation types had to be updated to allow using MessageResult in this component.

Test Plan:
Tested with the next diff - displayed it the search screen and checked that the messages are fetched and displayed correctly. Checked that the "End of results" and "No results..." messages display correctly. Checked the case when
the first fetch returns all results as well as the case when the user needs to scroll to fetch more messages a couple of times.

Reviewers: kuba, kamil, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7774