[native] Introduce RegistrationButtonContainer
Summary:
What we really want is the `gap` style [introduced in React Native 0.71](https://reactnative.dev/blog/2023/01/12/version-071#simplifying-layouts-with-flexbox-gap).

To make it so we have a consistent 16px above, between, and below buttons, my approach in this diff is:

1. Add a `RegistrationButtonContainer` that always applies an 8px margin above and below buttons.
2. Make all `RegistrationButton`s apply an additional 8px margin above and below themselves.

Depends on D7831

Test Plan:
With an artificially tall `ConnectEthereum` screen:

| before | after |
| {F540632} | {F540633} |

Other screens (which only have one button) are unchanged.

Reviewers: ginsu, atul

Reviewed By: ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7832