[landing] fix mobile nav bar peeking into screen when quickly resizing window
Summary:
As I was doing more tests for the landing page, I noticed that when I quickly resized the window, the `MobileNav` component would sometimes peek from the top and be visible. To fix this I made the background color of the `MobileNav` component transparent whenever it was not active so it would not be visible.

Depends on D7870

Test Plan:
Please watch the demo videos below

Before:
{F545427}

After:
{F548611}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7871