[native] Factor out loading spinner from SIWEPanel
Summary:
We want to use `SIWEPanel` in the new registration flow, but it currently renders a large spinner with `flex: 1` while loading, which doesn't match the designs for the new registration flow.

To make `SIWEPanel` usable in both the old and new context, this diff factors out the spinner code into `FullscreenSIWEPanel`, making `SIWEPanel` more reusable.

Depends on D7840

Test Plan: Make sure the loading spinner still appears full-screen in the old experience

Reviewers: ginsu, atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7841