[native] Introduce RegistrationContainer to handle SafeAreaView
Summary:
Before this diff, we wrapped all registration screens with a `SafeAreaView` in `RegistrationNavigator`.

However, this meant that the `SIWEPanel` background had a sharp cutoff at the `SafeAreaView` bounds.

In order to allow the `SIWEPanel` to appear over the `SafeAreaView`, this diff moves responsibility for rendering `SafeAreaView` to the individual screens. This is achieved with a new `RegistrationContainer` component.

Depends on D7844

Test Plan:
| before | after |
| {F541608} | {F541609} |

Reviewers: ginsu, atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7845