[native] Add loading spinner to RegistrationButton loading variant
Summary:
`ConnectEthereum` is the first place we're actually using the `loading` variant; before this point, I hadn't really implemented it.

This diff implements the `loading` variant with a loading spinner that is absolutely-positioned over the normal content.

The reason I took that approach is to avoid issues with the height changing. This way I don't have to worry about calculating the height of the text, or forcing the height of the text to be something specific.

Depends on D7845

Test Plan: {F541666}

Reviewers: ginsu, atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7846