[web] Convert ids during state check
Summary:
Because the ids will be different on the client and on the server we need to convert them during state sync. We convert them before hashing on the client side

Depends on D7851

Test Plan:
Enable id conversion and load a fresh (migration are not yet implemented) web app. Log `invalidKeys` in `processClientResponses` in `session-utils.js`.
- without the state check conversion there are invalid keys displayed
- with the state check conversion there are no errors

Reviewers: kamil, tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7852