[landing] clean up header icons user experience
Summary:
After some feedback from D7835 regarding the hover effect of the hamburger menu icon, I created this diff to address that feedback and add the hover effect in for the hamburger menu icon only for non-mobile/non-touch devices. In addition to adding the hover effect to the hamburger menu icon, I cleaned up the hover effects for the other icons in the header on mobile/touch devices. While this is not the most necessary thing, since we are immediately redirecting the user to a different webpage after the user presses on an icon, I figured it would be good to do to keep the styling consistent among the icons in the header

Depends on D7839

Test Plan:
Please watch the demo video to see how the header icons look and behave:

Using a non-mobile/non-touch device:
{F544529}

Using a mobile/touch device:
{F544540}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7863