[native] Introduce simple error handling
Summary:
This is a simplified version of error handling, without useful error message. Created a task https://linear.app/comm/issue/ENG-3916/better-error-handling to track implementing the proper solution.

Depends on D7883

Test Plan: Check if creating a link with white char or one that is already in use results in an error.

Reviewers: bartek, kamil, inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7884