[landing] show close icon when mobile nav is active
Summary:
This is the last diff that updates the header component. This diff changes the hamburger menu icon to a close icon depending if the mobile nav is active or not.

Depends on D7868

Test Plan:
Please see the screenshots below:

When the mobile nav is not active:
{F545332}

When the mobile nav is active:
{F545333}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7869