[lib] Create a store in redux for the invite links
Summary:
A couple of screens require invite links info, and it wasn't an issue until I added an option to update a link - then the data displayed on the screens was inconsistent. The approach with redux should be used from the beginning.

Depends on D7848

Test Plan: Tested in combination with the rest of the stack: updating a link on one screen and opening another screen where that link is displayed should show the current value.

Reviewers: bartek, kamil, inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7877