[native] Make panelSecondaryForeground light mode correspond to dark mode
Summary:
We're using `shadesDark80` and `shadesDark70` for dark mode, so I figure it made sense to use the corresponding `shadesWhite80` and `shadesWhite70` for light mode.

I also think this looks better. Note that `panelSecondaryForeground` is only used in the calendar, as well as the inner tile in `ExistingEthereumAccount`. Meanwhile, `panelSecondaryForegroundBorder` is only used in `RegistrationTile`. Screenshots for all three are in the Test Plan.

Depends on D7887

Test Plan: [screenshots]

Reviewers: ted

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D7888