[native] Add explicit borderColor for non-selected RegistrationTile
Summary:
This wasn't noticeable in dark mode, but is much more noticeable in light mode. Since we set a `borderWidth`, React Native color sets a border. I'm not sure how it determines a default color for that border, but I think it just picks black.

Depends on D7888

Test Plan: [screenshots]

Reviewers: ginsu, atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7889