[native] Add explicit borderColor for non-selected RegistrationTile
Summary:
This wasn't noticeable in dark mode, but is much more noticeable in light mode. Since we set a `borderWidth`, React Native color sets a border. I'm not sure how it determines a default color for that border, but I think it just picks black.

Depends on D7888

Test Plan:
| | before | after |
| light mode (not accessible; just showing for context) | {F546279} | {F546280} |
| dark mode (looks mostly the same) | {F546309} | {F546310} |

The only change in dark mode is that the tile looks slightly larger in the "after" version, since we're using the `backgroundColor` of the tile as the `borderColor`, instead of using a more black `backgroundColor` (which fades into the background).

Reviewers: ginsu, atul, ted

Reviewed By: atul, ted

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7889