[native] Introduce theme reducer to reset theme on log out
Summary:
I noticed that if the user sets a light theme, then logs out, and then enters the new registration flow, it will present itself in light mode.

We aren't prepared to handle a light mode registration flow. Perhaps the most complicated part is that we load a web view for Ethereum log in, and to support light mode there would require substantial work, including some way to communicate to the web view that we are in light mode.

Instead of adding support for light mode, this diff makes sure that we always reset the theme configuration on log out. (It also resets it on log in, to follow existing conventions in the codebase; for an example, see `lib/reducers/services-access-token-reducer.js`.)

Depends on D7888

Test Plan: Set light mode, then log out and enter the new registration flow. It should show in dark mode

Reviewers: ginsu, atul

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7891