[web] Introduce `CommunityCreationKeyserverLabel`
Summary:
Pulled common keyserver label out of `CommunityCreationModal` and `CommunityCreationMembersModal`.

> Why not make it more general and put it in `components`?

It has a hardcoded keyserver name (ashoat) so I think it makes sense to keep its use as limited as possible. `CommunityCreationKeyserverLabel` could be the basis for a more general purpose `KeyserverLabel` component in the future.

Test Plan: Things continue to look as expected w/ `CommunityCreationModal` and `CommunityCreationMembersModal`.

Reviewers: ashoat, ginsu

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7894