[web] Add placeholder prop to encrypted media, fix spinner
Summary:
This fixes the spinner CSS in modals for encrypted media. Without this change, it was displayed at the top instead of the center.
Also added props to the EncryptedMedia component to allow for a placeholder to be displayed when the media is being decrypted and to customize the style prop of the img / video elements. This will be used in later diffs.

Depends on D7899.

Test Plan: Ensured the spinner is centered in both chat and modal.

Reviewers: ashoat, atul, ginsu, inka

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7900