[web] Add thumbhash to encrypted media, fix spinner
Summary:
Also added props to the EncryptedMedia component to allow for a placeholder to be displayed when the media is being decrypted and to customize the style prop of the img / video elements. For videos, this uses the `LoadableVideo` component introduced in D7946

This also fixes the spinner CSS in modals for encrypted media. Without this change, it was displayed at the top instead of the center.

Depends on D7947.

Test Plan: Ensured the spinner is centered in both chat and modal.

Reviewers: ashoat, atul, ginsu, inka

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7900