[landing] change text content on HeroContent component
Summary:
As part of the landing redesign we are completely redesigning our Hero section. This diff updates the header and subheader of the hero section. Subsequent diffs will handle the Hero image, the email form, and the layout of the hero

Here is the Figma design for reference:
{F551181}

Depends on D7898

Test Plan:
Please see the screenshots below:

Desktop:
{F551206}

Tablet:
{F551207}

Phone:
{F551208}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7911