[landing] redesign the subscription form component
Summary:
As part of the Hero redesign for the landing page we need to redesign our `SubscriptionForm` component. The focus of this diff is to change the language and visuals (color and layout by separating the input field from the button) shown on `SubscriptionForm`. Here is a screenshot of the Figma doc (please note that a subsequent diff will handle displaying the feedback text shown in the Figma)

{F551448}

{F551449}

Depends on D7914

Test Plan:
Please see the screenshots below:

Desktop Normal:
{F551450}

Hover:
{F551451}

Success Case:
{F551453}

Error Case:
{F551454}

Mobile View:
{F566246}

Tablet View:
{F566247}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7915