[landing] redesign hero section layout
Summary:
This diff redesigns the hero section to match the new designs shown in the figma document. In this diff we replaced the hero image and arranged the elements to match the designs. For the hero image, I did remove the purple blur background effect to make positioning the image easier; however, a subsequent diff will add those blur effects back in. For context here are the designs:

{F569416}

{F569415}

Depends on D8055

Test Plan:
Please watch the demo video and see the screenshots below:

Desktop/Responsive behavior:
{F569624}

Tablet:
{F569425}

Mobile:
{F569426}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek

Differential Revision: https://phab.comm.dev/D8075