[keyserver] Don't use web version when sending desktop notifs
Summary:
For windows and macos we don't want to use the web's `codeVersion` when sending notifs. This diff keeps the previous behaviour of keeping the `codeVersion` as `-1` in case of desktop notifs. Made a task for handling desktop versions on the keyserver: [ENG-3917](https://linear.app/comm/issue/ENG-3917)

Depends on D7855

Test Plan:
Set web version, try sending a macOS notification. Check if the `codeVersion` in the `notifications` table is `-1`.

Reviewers: kamil, tomek

Subscribers: