[lib] Add reducer and endpoint calling code
Summary:
The reducer deletes an entry in the store. In the future we will have to do more here, as not only public links can be disabled.

Depends on D7906

Test Plan: Confirm disabling a link and check if it was deleted, e.g. by verifying that disable link button disappears.

Reviewers: bartek, kamil, inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7907