[native] Introduce disable link button
Summary:
The button is visible only when a link is present.
Clicking the button results in alert being shown where a user is asked to confirm the action.
Confirming results in the link being disabled.

Depends on D7907

Test Plan:
Check if the button appears only when a link is present.
Check if clicking it results in an alert.
Check if canceling doesn't disable the link.
Check if confirming disables the link.

Reviewers: bartek, kamil, inka

Subscribers: ashoat, atul, ted

Differential Revision: https://phab.comm.dev/D7908