[native] Extract onClose-disabling behavior from SIWEPanel to FullscreenSIWEPanel
Summary:
In the original `SIWEPanel`, we [implemented logic](https://phab.comm.dev/D6068) to make sure that after a successful SIWE, we didn't trigger `goBackToPrompt` after the closing of the bottom sheet.

However, this logic is currently messing with the new implementation, where it's preventing a second SIWE authorization since the first one was never cleared since the `onClose` wasn't called.

This diff moves the behavior that blocks `goBackToPrompt` to `FullscreenSIWEPanel`, so that both use cases work correctly.

Depends on D7927

Test Plan:
1. I tested the old workflow and made sure that `goBackToPrompt` wasn't triggered after a successful SIWE
2. See videos below for the impact on the new workflow:

[videos]

Reviewers: atul, ginsu

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7928