[web] Add expanded and toggleExpanded to CommunityDrawerItemHandler type
Summary:
issue: https://linear.app/comm/issue/ENG-3176/make-pressing-a-drawer-item-in-chat-tab-filter-available-chats
Adding to handlers for drawer items `expanded` state. This is now kept by the wrappe of `CommunityDrawerItem` - `CommunityDrawerItemChat`. But since we need `expanded` to be handled by
`CommunityDrawerItemCommunityHandler` for `CommunityDrawerItemCommunity`, we will handle it in the handler for `CommunityDrawerItemChat` as well, for better code clarity. These fileds will ne
used in the next diff.

Test Plan: texted with the next diff

Reviewers: kamil, tomek, michal

Reviewed By: tomek

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D6975