Remove ListLoadingIndicator background color
Summary:
issue: https://linear.app/comm/issue/ENG-3880/fix-spinner-color-not-matching
in all cases we want the spinner colour to match the backgroud behind it. The colour doen't need to be specified explicitly

Test Plan: Checked all screens that use `ListLoadingIndicator` - checked that they look the same, and on the search screen the loader background matches the screen background. Checked light and dark mode

Reviewers: kamil, kuba, bartek

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7935