[web] Support thumbhash in modals
Summary:
This adds support for thumb hash in modals. Unfortunately, CSS in encrypted media does not know its width and height so had to do measuring magic to get it to work with thumbhash, before actual `src` is available. Encrypted video support is added in D7900

Test Plan: Throttled network, slowed down decryption. Ensured the loading behavior is the same as for in-chat media (both non-encrypted images and videos). Resized browser window. Ensured the aspect ratio is maintained.

Reviewers: ashoat, atul, ginsu, inka

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7947