[landing] introduce logo svg assets for competitor comparison
Summary:
Based on some feedback received in D7905 and some discussion in the design channel we decided that a better way to get the competitor logo assets would be to bundle the svgs directly instead of fetching them over the network. To get the svgs I downloaded them straight from the Figma file and copy and pasted the svg contents into its own react component. In a subsequent diff I will make some modifications to each component to allow for a `size` prop so that we can modify the size of the logos depending on its usage.

{F554436}

Test Plan:
Rendered the logos on the landing page

{F554382}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ted, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7943