[native] Fix reply not working in search results
Summary:
issue: https://linear.app/comm/issue/ENG-3881/reply-doesnt-work-in-search-screen-and-in-pinned-messages-screen
Reply from within the search screen didn't work. We need to navigate back to the thread to be able to reply.
Adding `MessageSearchRouteName` to `threadRoutes` is necessary, because we don't want chat input bar to remove its listener for input state events, which happens if `activeThreadSelector` returns null or
an id of a different thread (see `isActive` variable in chat-input-bar.react.js). `activeThreadSelector` returns null if `threadRoutes` doesn't contain the current route name.
`ThreadSettingsRouteName` is also in `threadRoutes`, and the search screen is used in a similar fashion to thread settings screen.

Test Plan: Pressed "reply" on the search screen. Checked that the correct thread was opened and the input bar message was prepended with the reply.

Reviewers: kamil, kuba, bartek, ashoat

Subscribers: tomek, atul

Differential Revision: https://phab.comm.dev/D7942