[server] Handle web invite links on keyserver
Summary:
This is a really simple diff that achieves a lot.
After clicking an invite link on web a user is redirected to the app. In the app, this new route isn't implemented, so the user either sees a login screen or a place which was opened in the most recent session, which is exactly what we want.

Instead of simply using `${baseRoutePath}invite/${secret}` I had to create something different, because in case of an empty base path, we would get a redirect loop.

Test Plan:
Open `localhost:3000/invite/abc` on web - it should be redirected to `localhost:3000/comm/handle/invite/abc` and then e.g. to `localhost:3000/comm/`.
On iOS accepting invite flow should still work.

Reviewers: bartek, kamil, inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7948