[native] Introduce `CommunityCreationNavigator`
Summary:
Heavily inspired by `RegistrationNavigator`, this diff introduces `CommunityCreationNavigator`. This navigator will "handle" the screens of the community creation flow. Specifically the community configuration screen and the optional "Add members" screen.

WARNING: Please let me know if this is the wrong approach. My understanding is that because we have > 1 screen in sequence as part of our flow we should create a new navigation stack. HOWEVER, I have a very very limited understanding of navigation on `native`, so please correct me if that assumption is incorrect and I should approach this differently.

Test Plan:
1. Set `isCommunityCreationButtonEnabled` to `true`.
2. Open community drawer and press "Create community."
3. Observe that `CommunityConfiguration` is displayed (by looking at screen "title").

Reviewers:
ashoat, ginsu