[web] Open invite link modal when the link is clicked
Summary:
When a user isn't logged in, we show login panel. When a user logs in, we show a modal.

Depends on D7958

Test Plan:
Open invite link as a logged-in user - a modal should appear.
Open the link as a logged-out user. A login form should appear. After successful login, the modal appears.

Reviewers: bartek, kamil, inka

Reviewed By: inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7959