[web] Call join thread endpoint
Summary:
Clicking join button results in calling server endpoint. Depending on the result either user sees a modal with an error, or a modal is closed.

Depends on D7957

Test Plan:
Open invite link, accept the invitation - modal should be closed and the community should appear on the list.
Modify join thread endpoint so that it always throws. Accept the invitation - a message about invalid link should appear.

Reviewers: bartek, kamil, inka

Reviewed By: inka

Subscribers: ashoat, atul

Differential Revision: https://phab.comm.dev/D7958