[native] Use search users in chat composer
Summary:
Uses search users endpoint in native chat composer. To handle that we might now have to have these users in the userStore now `user-list` returns whole user info instead of just an id. In other places where this is used I just extract the user id.

Depends on D8009

Test Plan: Check if when searching in chat composer, there are users not in userStore displayed. Try clicking them and check if they are added correctly. Try selecting someone else and then selecting someone not in user store, and check if the "you have to friend them first" alert appears.

Reviewers: kamil, inka

Reviewed By: kamil

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D8010