[web] Don't hide pinned messages on reload
Summary:
issue: https://linear.app/comm/issue/ENG-3896/message-pin-view-on-web-refreshes-when-user-navigates-back-and-away
When the user navigates to a different tab / app the `connectionStatus` is changed to `disconnected`. This indirectly causes the `callFetchPinnedMessages` variable in `MessageResultsModal` to change when the user comes back to comm, which
in turn calls the Effect hook that fetches pinned messages.
Previously in such case we stopped showing the old results, and show only the loading indicator. Now we show the old results and the loading indicator above them, and once the fetch is completed the results are updated.
Some changes to css were made to stop the modal from resizing when the loading indicator appears / disappears (resulting in jumping screen effect).

Test Plan:
{F558835}

{F558836}

{F558837}

{F558838}

Reviewers: kuba, michal, ginsu, rohan

Reviewed By: rohan

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D7981