[web] Update wordmark positioning
Summary:
Context in [ENG-3948](https://linear.app/comm/issue/ENG-3948/comm-logo-is-horizontally-offset-on-web-in-prod). This was initially set in D2923, and looked properly aligned at the time. But the design of the web app has changed since then, and it no longer looks right.

This diff updates the `padding-left` to 34px to align with where the avatar for a community appears in the community navigation drawer.

Test Plan: {F559902}

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7999