[native] Delay PasswordSelection autoFocus on Android
Summary:
While playing around with 1Password on Android, I noticed that the autofill prompt didn't appear when we `autoFocus`ed the password `TextInput` in `PasswordSelection`. However, it would appear if the user subsequently pressed / selected the `TextInput` manually.

Test Plan:
I tried various delays for the focus on two phones: my Pixel 6a, and the older Pixel 3 in the office. 250ms consistently worked for both, while 100ms did not.

Reviewers: atul, ginsu

Subscribers: