[native] Set RegistrationTextInput keyboardAppearance based on theme
Summary:
For the registration flow, this will force dark mode since the registration flow can only be triggered in dark mode.

For other uses of `RegistrationTextInput`, this will set the `keyboardAppearance` to match the rest of the app's theme.

Depends on D8000

Test Plan: Tested to make sure that the `keyboardAppearance` was getting set correctly

Reviewers: atul, ginsu

Reviewed By: atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8001