[web] Change navInfo.selectedUserList to contains user infos
Summary:
The users we get from search users endpoint might not be in the userStore. Because of that this diff changes the navInfo to contain the whole userInfos instead of just ids. When selecting a new user in the chat composer we save user info in the navInfo. During redux state updates we use the previous navInfo as the source of the userInfos in the `navInfoFromURL`.

Depends on D8009

Test Plan: Try adding and removing users from chat composer. Check if the url and navInfo are updated correctly. Try navigating to `chat/thread/new/<non-existant-user-id>` and check if it is removed.

Reviewers: kamil, inka

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D8011