[native] Replace generic useSaveUserAvatar with new generic setUserAvatar
Summary:
In the last diff, I introduced a generic `setUserAvatar` in `EditUserAvatarProvider`. In a later diff, I'll add some logic that allows me to "override" the action that gets performed for a successful avatar selection, so that I can reuse it in the registration flow.

This code replaces `useSaveUserAvatar` with the generic utility, to make the codebase more consistent.

Depends on D8017

Test Plan: Test setting emoji avatar for user from iOS simulator

Reviewers: ginsu, atul

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D8018