[native] Replace removeThreadAvatar with generic setThreadAvatar
Summary: Decided to do the same thing for thread avatars here as I did for user avatars in D8017. It's good to unify codepaths, and this diff will allow me to deprecate `useSaveThreadAvatar` in the next diff, replacing it with the shared utilities in `EditThreadAvatarProvider`.

Depends on D8019

Test Plan: Test clearing thread avatar via iOS simulator

Reviewers: ginsu, atul

Subscribers: