[landing] introduce competitor feature base component
Summary:
The competitor comparison cards and competitor comparison modal will have a similar structure so I factored out the base component so that both places can use this code. A subsequent diff will handle the introduction of the competitor comparison modal. This diff is mainly a move diff

Depends on D7905

Test Plan: The competitor feature cards look and behave the same, and I will do more testing of the competitor comparison modal when I introduce it

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D8024