[landing] introduce competitor comparison
Summary:
This diff introduces the Competitor comparison component. This is the first part of introducing the complete competitor comparison. This diff introduces the component and renders the header as well as the `competitorSelector`.  The next diff will introduce the competitor cards to this component.

Designs:
{F566169}

Depends on D8033

Test Plan:
Please see the screenshots/demo videos below to see how everything looks/behaves:

Desktop/Responsive/Behavior
{F566170}

Tablet:
{F566176}

Mobile:
{F566178}

Reviewers: atul, kamil

Reviewed By: atul

Subscribers: ted, ashoat, tomek, atul

Differential Revision: https://phab.comm.dev/D8034