[native] Make EmojiThreadAvatarCreation work with activeThreadSelector
Summary: In D8021, we changing how our loading status worked. Before that the custom key name was not thread specific, but afterwards it is.

In the new system, for `EditThreadAvatarProvider` to know which `threadID` to look up, it calls `activeThreadSelector` on the nav state. For this to work, we need two things to be true:

1. The React Navigation route needs to be in `routeNames` in `route-names.js`
2. The React Navigation route params should have a field called `threadInfo`

This diff makes both changes so that the loading detection works

Depends on D8022

Test Plan: I added an artificial delay on the keyserver side and confirmed that the loading spinner appears currently now (which was a regression introduced in D8021)

Reviewers: ginsu, atul

Subscribers: